Addressed issues partially :scroll:
- Moved CE discussion within tce-core folder
- Extract activity-discussion wrapper in common
- Removed unecessary mixin
- Styles cleanup :lipstick:
