+++ client/store/modules/repository/comments/mutations.js
@@ -8,11 +8,11 @@ import {
   update
 } from '@/store/helpers/mutations';
 
-const markSeenComments = (state, { activityUid, ceUid, lastCommentAt }) => {
-  const key = ceUid ? 'contentElement' : 'activity';
+const markSeenComments = (state, { activityUid, elementUid, lastCommentAt }) => {
+  const key = elementUid ? 'contentElement' : 'activity';
   state.seen[key] = {
     ...state.seen[key],
-    [ceUid || activityUid]: lastCommentAt
+    [elementUid || activityUid]: lastCommentAt
   };
 };
